Changes kernel API to use multi-dimensional fixed size C arrays
Previously a custom templated type with overloaded subscript operator was
used. This is changed in favor of plain C arrays to enable compatibility to
the Tepanade AD framework. As a consequence the order of indices is reversed:
the index into the vector mapping comes first (if present), followed by
indices into the current dataset element. Binary compatibility to the Fortran
kernel API is maintained.

In order to keep the type checking in place, the extents have to be stripped
from the kernel parameter to get the scalar primitive datatype of the argument
to check against. This is done with help of the RemoveExtent struct.